Client/requirements.txt
Flask
PyMySQL
bcrypt
PyModbus
PySerial